+++ platforms/gradle/foundry-gradle-plugin/src/main/kotlin/foundry/gradle/FoundryExtension.kt
@@ -105,7 +105,7 @@ constructor(
   internal fun applyTo(project: Project) {
     val logVerbose = foundryProperties.foundryExtensionVerbose
     // Dirty but necessary since the extension isn't configured yet when we call this
-    project.afterEvaluate {
+    project.afterEvaluateSafe {
       featuresHandler.applyTo(project)
 
       var kaptRequired = false